Move print function and its tests to a separate file.
