+++ internal/game/game.go
@@ -3,8 +3,6 @@ package game
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
 	"os"
 )
 
@@ -84,27 +82,5 @@ func InitBoard(filePath *string) (*BoardState, error) {
 	} else {
 		// No file path, return the default board.
 		return &DefaultBoard, nil
-	}
-}
-
-// Print the game board to the console.
-func (board *BoardState) Print(writer io.Writer) {
-	fmt.Fprintln(writer, "    1    2    3    4    5  ")
-	fmt.Fprintln(writer, ". +----+----+----+----+----+")
-	for rowIndex, row := range board.Board {
-		fmt.Fprintf(writer, "%c ", rune(int('a')+rowIndex))
-		fmt.Fprint(writer, "|")
-		for _, cell := range row {
-			if cell == 1 {
-				fmt.Fprint(writer, " 🟢 ")
-			} else if cell == 2 {
-				fmt.Fprint(writer, " 🔴 ")
-			} else {
-				fmt.Fprint(writer, "    ")
-			}
-			fmt.Fprint(writer, "|")
-		}
-		fmt.Fprintln(writer)
-		fmt.Fprintln(writer, ". +----+----+----+----+----+")
 	}
 }
