+++ internal/game/game_test.go
@@ -1,7 +1,6 @@
 package game
 
 import (
-	"bytes"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -57,50 +56,4 @@ func TestInitBoardWithFilePath(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, expected, board, "should be an ongoing game board")
-}
-
-func TestBoardPrinting(t *testing.T) {
-	expected := `    1    2    3    4    5  
-. +----+----+----+----+----+
-a | 🟢 | 🟢 | 🟢 |    |    |
-. +----+----+----+----+----+
-b | 🟢 | 🟢 |    |    |    |
-. +----+----+----+----+----+
-c | 🟢 |    |    |    | 🔴 |
-. +----+----+----+----+----+
-d |    |    |    | 🔴 | 🔴 |
-. +----+----+----+----+----+
-e |    |    | 🔴 | 🔴 | 🔴 |
-. +----+----+----+----+----+
-`
-
-	var output bytes.Buffer
-	DefaultBoard.Print(&output)
-
-	assert.Equal(t, expected, output.String(), "should have printed a default board")
-}
-
-func TestOngoingGameBoardPrinting(t *testing.T) {
-	expected := `    1    2    3    4    5  
-. +----+----+----+----+----+
-a |    | 🟢 | 🟢 |    |    |
-. +----+----+----+----+----+
-b |    | 🟢 |    |    |    |
-. +----+----+----+----+----+
-c | 🟢 | 🟢 |    | 🔴 |    |
-. +----+----+----+----+----+
-d | 🟢 |    | 🔴 | 🔴 | 🔴 |
-. +----+----+----+----+----+
-e |    |    |    | 🔴 | 🔴 |
-. +----+----+----+----+----+
-`
-
-	path := ongoingGameStateTestPath
-	board, err := InitBoard(&path)
-
-	var output bytes.Buffer
-	board.Print(&output)
-
-	assert.Nil(t, err)
-	assert.Equal(t, expected, output.String(), "should have printed an ongoing game board")
 }
